requirements.txt
pandas
requests
google-search-results
python-dotenv